Convert resolvers to folder
