+++ services/src/modules/registry-schema/resolvers/update-resource-group.ts
@@ -1,0 +1,46 @@
+import pLimit from 'p-limit';
+import * as _ from 'lodash';
+import { createSchemaConfig } from '../../graphql-service';
+import { applyResourceGroupUpdates } from '../../resource-repository';
+import { validateResourceGroupOrThrow } from '../../validation';
+import { transformResourceGroup as applyPluginForResourceGroup } from '../../plugins';
+import PolicyAttachmentsGenerator from '../policy-attachments-generator';
+import resourceRepository from '../repository';
+import { ResourceGroupInput } from '../types';
+
+const singleton = pLimit(1);
+
+export default async function (updates: ResourceGroupInput, dryRun = false) {
+  return singleton(async () => {
+    const policyAttachments = new PolicyAttachmentsGenerator();
+
+    try {
+      const { resourceGroup } = await resourceRepository.fetchLatest();
+      const existingPolicies = _.cloneDeep(resourceGroup.policies);
+
+      const newRg = applyResourceGroupUpdates(resourceGroup, updates);
+      const registryRg = _.cloneDeep(newRg);
+
+      const gatewayRg = await applyPluginForResourceGroup(newRg);
+      validateResourceGroupOrThrow(gatewayRg);
+      await createSchemaConfig(gatewayRg);
+
+      // Policy definitions generated by plugins will always be re-generated here. This is probably
+      // better than reading the previous gateway resource group just for this comparison
+      const policiesDiff = _.differenceWith(gatewayRg.policies, existingPolicies, _.isEqual);
+      await policyAttachments.generate(policiesDiff);
+
+      if (!dryRun) {
+        await policyAttachments.writeToRepo();
+        await Promise.all([
+          resourceRepository.update(registryRg, { registry: true }),
+          resourceRepository.update(gatewayRg),
+        ]);
+      }
+    } finally {
+      await policyAttachments.cleanup();
+    }
+
+    return { success: true };
+  });
+}
